Injected mouseToClip and saveToData url from SceneManager
